Send page widgets refactoring
